+++ assets/javascripts/mermaid_load.js
@@ -1,3 +1,5 @@
 $(function() {
-  mermaid.initialize({ startOnLoad: true });
+  mermaid.initialize({ startOnLoad: true,
+                       theme: 'default',
+                       themeVariables: { 'fontSize': '12px' } });
 });